Update documentation and rename laporan.py to reports.py
README.md has been revised for improved clarity, updated feature descriptions, and new module explanations. The helper file 'laporan.py' was renamed to 'reports.py' for consistency, and main.py was updated to import from the new filename.
