+++ helpers/reports.py
